+++ app/src/main/java/app/revanced/manager/ui/viewmodel/MainViewModel.kt
@@ -190,6 +190,7 @@ class MainViewModel(
             keystoreManager.import(
                 "alias",
                 settings.keystorePassword,
+                settings.keystorePassword,
                 keystoreBytes.inputStream()
             )
         }